Move wsclean predict into RIME package
